+++ examples/e2e.py
@@ -12,7 +12,6 @@ def main() -> None:
     memory_strategy = 'high'  # can be high or low depending on memory availability/GPU setup
     product_dst_dir = Path('../notebooks/los-angeles')  # directory to save the final products
     apply_water_mask = True  # apply water mask to the data
-    n_lookbacks = 3  # number of lookbacks to use for change confirmation within SAS
     water_mask_path = None  # Path('../notebooks/los-angeles/water_mask.tif')  # path to an existing water mask file
     device = 'cpu'  # can be cpu, cuda, mps or best
     n_workers_for_norm_param_estimation = 4  # number of workers for normal parameter estimation
@@ -29,7 +28,6 @@ def main() -> None:
         memory_strategy=memory_strategy,
         product_dst_dir=product_dst_dir,
         apply_water_mask=apply_water_mask,
-        n_lookbacks=n_lookbacks,
         water_mask_path=water_mask_path,
         device=device,
         n_workers_for_norm_param_estimation=n_workers_for_norm_param_estimation,