+++ examples/e2e.sh
@@ -8,6 +8,6 @@ dist-s1 run \
     --n_workers_for_norm_param_estimation 5 \
     --batch_size_for_despeckling 100 \
     --optimize false \
-    --confirmation_strategy 'compute_baseline' \
+    --confirmation 'false' \
     --apply_water_mask true \
     --water_mask_path '../notebooks/los-angeles/11SLT_water_mask.tif'