+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2025-07-07
+
+## Removed
+- `n_lookbacks` (which should be called `n_baselines`) for computing confirmation within the SAS
+- ...
+
+## Changed
+- Multiwindow strategy is now the default
+- Confirmation_strategy is now `confirmation` and simply a boolean
+
+## Fixed
+- All the tests with the updates above.
+
 ## [1.0.1] - 2025-06-05
 
 ### Changed